add DependencyInjection support.
