+++ Microsoft.Extensions.Caching.CSRedis.sln
@@ -1,0 +1,44 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27625.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{C3F4880A-561D-4CE8-BEB0-3A5A9C650650}"
+	ProjectSection(SolutionItems) = preProject
+		README.md = README.md
+	EndProjectSection
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{097A219F-A6E0-40DA-BF74-768A0D962FA4}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Extensions.Caching.CSRedis", "src\Microsoft.Extensions.Caching.CSRedis\Microsoft.Extensions.Caching.CSRedis.csproj", "{70A6AE57-AB75-4B07-83E9-99A6DBE1F694}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{1F1FA09E-D112-4599-8781-E81BDEC00094}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Extensions.Caching.CSRedis.Tests", "test\Microsoft.Extensions.Caching.CSRedis.Tests\Microsoft.Extensions.Caching.CSRedis.Tests.csproj", "{60EBFC97-63F3-48BA-B234-8D9E2285B022}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{70A6AE57-AB75-4B07-83E9-99A6DBE1F694}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{70A6AE57-AB75-4B07-83E9-99A6DBE1F694}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{70A6AE57-AB75-4B07-83E9-99A6DBE1F694}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{70A6AE57-AB75-4B07-83E9-99A6DBE1F694}.Release|Any CPU.Build.0 = Release|Any CPU
+		{60EBFC97-63F3-48BA-B234-8D9E2285B022}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{60EBFC97-63F3-48BA-B234-8D9E2285B022}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{60EBFC97-63F3-48BA-B234-8D9E2285B022}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{60EBFC97-63F3-48BA-B234-8D9E2285B022}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{70A6AE57-AB75-4B07-83E9-99A6DBE1F694} = {097A219F-A6E0-40DA-BF74-768A0D962FA4}
+		{60EBFC97-63F3-48BA-B234-8D9E2285B022} = {1F1FA09E-D112-4599-8781-E81BDEC00094}
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {CD487DE3-7066-4F0F-AB32-3E1A3BD02DA0}
+	EndGlobalSection
+EndGlobal
