+++ src/Microsoft.Extensions.Caching.CSRedis/Microsoft.Extensions.Caching.CSRedis.csproj
@@ -1,9 +1,9 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
-    <AssemblyName>Caching.CSRedis</AssemblyName>
-    <PackageId>Caching.CSRedis</PackageId>
-    <RootNamespace>Caching.CSRedis</RootNamespace>
+    <AssemblyName>Microsoft.Extensions.Caching.CSRedis</AssemblyName>
+    <PackageId>Microsoft.Extensions.Caching.CSRedis</PackageId>
+    <RootNamespace>Microsoft.Extensions.Caching.CSRedis</RootNamespace>
     <Version>2.3.9</Version>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <PackageProjectUrl>https://github.com/2881099/Microsoft.Extensions.Caching.CSRedis</PackageProjectUrl>
@@ -16,10 +16,8 @@
     <DocumentationFile>bin\Debug\netstandard2.0\Caching.CSRedis.xml</DocumentationFile>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Caching.Abstractions" Version="2.1.0" />
+    <PackageReference Include="CSRedisCore" Version="2.3.9" />
+    <PackageReference Include="Microsoft.Extensions.Caching.Redis" Version="2.1.0" />
     <PackageReference Include="Newtonsoft.Json" Version="11.0.2" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\csredis\CSRedisCore.csproj" />
   </ItemGroup>
 </Project>
